+++ app/dashboard/properties/[propertyId]/events/add/_components/ServiceWorkContent.tsx
@@ -4,10 +4,13 @@ import { EventServiceWorkSelector } from './Selectors/EventServiceWorkSelector';
 import { EventTargetSelector } from './Selectors/EventTargetSelector';
 
 export function ServiceWorkContent() {
+  const { eventData, refs } = useEventFormContext();
   return (
     <>
       <EventTargetSelector />
-      <EventServiceWorkSelector />
+      {eventData.target_id != getIdByLabel(refs.eventTargets, 'Muu') && (
+        <EventServiceWorkSelector />
+      )}
     </>
   );
 }